requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9
reportlab==4.0.7
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
